+++ packages/gamut/src/Tip/__tests__/InfoTip.test.tsx
@@ -1,12 +1,15 @@
 import { setupRtl } from '@codecademy/gamut-tests';
-import { act } from '@testing-library/react';
+import { act, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { createRef } from 'react';
 
+import { Anchor } from '../../Anchor';
+import { Text } from '../../Typography';
 import { InfoTip } from '../InfoTip';
 
 const info = 'I am information';
 const renderView = setupRtl(InfoTip, {
-  info: 'I am information',
+  info,
 });
 
 describe('InfoTip', () => {
@@ -45,6 +48,66 @@ describe('InfoTip', () => {
 
       // The first get by text result is the a11y text, the second is the actual tip text
       expect(view.queryAllByText(info).length).toBe(2);
+    });
+
+    it('closes the tip when Escape key is pressed and returns focus to the button', async () => {
+      const { view } = renderView({
+        placement: 'floating',
+      });
+
+      const button = view.getByLabelText('Show information');
+      await act(async () => {
+        await userEvent.click(button);
+      });
+
+      expect(view.queryAllByText(info).length).toBe(2);
+
+      await act(async () => {
+        await userEvent.keyboard('{Escape}');
+      });
+
+      await waitFor(() => {
+        expect(view.queryByText(info)).toBeNull();
+      });
+      expect(button).toHaveFocus();
+    });
+
+    it('closes the tip with links when Escape key is pressed and returns focus to the button', async () => {
+      const linkText = 'cool link';
+      const linkRef = createRef<HTMLAnchorElement>();
+      const { view } = renderView({
+        placement: 'floating',
+        info: (
+          <Text>
+            Hey! Here is a{' '}
+            <Anchor href="https://giphy.com/search/nichijou" ref={linkRef}>
+              {linkText}
+            </Anchor>{' '}
+            that is super important.
+          </Text>
+        ),
+        onClick: ({ isTipHidden }: { isTipHidden: boolean }) => {
+          if (!isTipHidden) {
+            linkRef.current?.focus();
+          }
+        },
+      });
+
+      const button = view.getByLabelText('Show information');
+      await act(async () => {
+        await userEvent.click(button);
+      });
+
+      expect(view.queryAllByText(linkText).length).toBe(2);
+
+      await act(async () => {
+        await userEvent.keyboard('{Escape}');
+      });
+
+      await waitFor(() => {
+        expect(view.queryByText(linkText)).toBeNull();
+      });
+      expect(button).toHaveFocus();
     });
   });
 });