+++ packages/gamut/src/Tip/shared/types.tsx
@@ -78,6 +78,7 @@ export type TipPlacementComponentProps = Omit<
   escapeKeyPressHandler?: (event: React.KeyboardEvent<HTMLDivElement>) => void;
   id?: string;
   isTipHidden?: boolean;
+  popoverContentRef?: React.RefObject<HTMLDivElement>;
   type: 'info' | 'tool' | 'preview';
   wrapperRef?: React.RefObject<HTMLDivElement>;
   zIndex?: number;